requirements.txt
blinker==1.9.0
click==8.1.8
exceptiongroup==1.3.1
Flask==3.1.3
importlib_metadata==8.7.1
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
markdown-it-py==3.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
packaging==26.0
pillow==11.3.0
pluggy==1.6.0
Pygments==2.19.2
pytest==8.4.2
rich==14.3.3
tomli==2.4.0
typing_extensions==4.15.0
Werkzeug==3.1.6
zipp==3.23.0
